Move chatbot provider module
